+++ shared/src/commonMain/kotlin/com/turbosokol/mypurchases/common/purchases/redux/PurchaseState.kt
@@ -6,8 +6,8 @@ import comturbosokolmypurchases.PurchaseDb
 
 data class PurchaseState(
     val progress: Boolean,
-    val purchaseId: Long,
-    val coast: Long,
+    val purchaseId: Double,
+    val coast: Double,
     val description: String? = null,
     val purchaseItems: List<PurchaseDb>,
     val editablePurchase: PurchaseDb
@@ -17,16 +17,16 @@ data class PurchaseState(
         fun getDefault(): PurchaseState = PurchaseState(
             progress = false,
             purchaseItems = emptyList(),
-            purchaseId = 0,
-            coast = 0,
+            purchaseId = 0.0,
+            coast = 0.0,
             description = null,
-            editablePurchase = PurchaseDb(id = 0L, parent = "", coast = 0, title = "")
+            editablePurchase = PurchaseDb(id = 0L, parent = "", coast = 0.0, title = "")
         )
     }
 }
 
 sealed class PurchaseAction: Action {
-    data class AddPurchase(val parentTitle: String, val coast: Long, val description: String? = null): PurchaseAction()
+    data class AddPurchase(val parentTitle: String, val coast: Double, val description: String? = null): PurchaseAction()
     object GetAllPurchases: PurchaseAction()
     data class GetAllPurchasesByParent(val parentTitle: String): PurchaseAction()
     data class SetPurchases(val purchaseItems: List<PurchaseDb>): PurchaseAction()
